Cleanup run.go
